Distributed attributes implementation is based now on live nodes

+++ src/main/java/io/vlingo/xoom/cluster/model/attribute/RemoteAttributeRequestHandler.java
@@ -7,54 +7,81 @@
 
 package io.vlingo.xoom.cluster.model.attribute;
 
+import io.vlingo.xoom.actors.Logger;
 import io.vlingo.xoom.cluster.model.attribute.message.ReceivedAttributeMessage;
-import io.vlingo.xoom.wire.node.Configuration;
+import io.vlingo.xoom.wire.node.Id;
+import io.vlingo.xoom.wire.node.Node;
+
+import java.util.function.Function;
 
 final class RemoteAttributeRequestHandler {
-  private final Configuration configuration;
   private final ConfirmingDistributor confirmingDistributor;
   private final AttributeSetRepository repository;
+  private final Function<Id, Node> nodeLookup;
+  private final Logger logger;
   
-  RemoteAttributeRequestHandler(final ConfirmingDistributor confirmingDistributor, final Configuration configuration, final AttributeSetRepository repository) {
+  RemoteAttributeRequestHandler(final ConfirmingDistributor confirmingDistributor, final AttributeSetRepository repository,
+                                final Function<Id, Node> nodeLookup, final Logger logger) {
     this.confirmingDistributor = confirmingDistributor;
-    this.configuration = configuration;
     this.repository = repository;
+    this.nodeLookup = nodeLookup;
+    this.logger = logger;
   }
 
   void addAttribute(final ReceivedAttributeMessage request) {
-    AttributeSet attributeSet = repository.attributeSetOf(request.attributeSetName());
-    if (attributeSet.isNone()) {
-      attributeSet = AttributeSet.named(request.attributeSetName());
-      repository.add(attributeSet);
+    final Node sourceNode = nodeLookup.apply(request.sourceNodeId());
+    if (sourceNode == null) {
+      logger.warn("Failed to addAttribute because source node " + request.sourceNodeId() + " is not part of the cluster anymore!");
+    } else {
+      AttributeSet attributeSet = repository.attributeSetOf(request.attributeSetName());
+      if (attributeSet.isNone()) {
+        attributeSet = AttributeSet.named(request.attributeSetName());
+        repository.add(attributeSet);
+      }
+      final TrackedAttribute tracked = attributeSet.addIfAbsent(request.attribute());
+      confirmingDistributor.confirm(request.trackingId(), attributeSet, tracked, request.type(), sourceNode);
     }
-    final TrackedAttribute tracked = attributeSet.addIfAbsent(request.attribute());
-    confirmingDistributor.confirm(request.trackingId(), attributeSet, tracked, request.type(), configuration.nodeMatching(request.sourceNodeId()));
   }
 
   void createAttributeSet(final ReceivedAttributeMessage request) {
-    AttributeSet attributeSet = repository.attributeSetOf(request.attributeSetName());
-    if (attributeSet.isNone()) {
-      attributeSet = AttributeSet.named(request.attributeSetName());
-      repository.add(attributeSet);
+    final Node sourceNode = nodeLookup.apply(request.sourceNodeId());
+    if (sourceNode == null) {
+      logger.warn("Failed to createAttributeSet because source node " + request.sourceNodeId() + " is not part of the cluster anymore!");
+    } else {
+      AttributeSet attributeSet = repository.attributeSetOf(request.attributeSetName());
+      if (attributeSet.isNone()) {
+        attributeSet = AttributeSet.named(request.attributeSetName());
+        repository.add(attributeSet);
+      }
+      confirmingDistributor.confirmCreate(request.trackingId(), attributeSet, sourceNode);
     }
-    confirmingDistributor.confirmCreate(request.trackingId(), attributeSet, configuration.nodeMatching(request.sourceNodeId()));
   }
 
   void removeAttributeSet(final ReceivedAttributeMessage request) {
-    AttributeSet attributeSet = repository.attributeSetOf(request.attributeSetName());
-    if (attributeSet.isDefined()) {
-      attributeSet = AttributeSet.named(request.attributeSetName());
-      repository.remove(request.attributeSetName());
+    final Node sourceNode = nodeLookup.apply(request.sourceNodeId());
+    if (sourceNode == null) {
+      logger.warn("Failed to removeAttributeSet because source node " + request.sourceNodeId() + " is not part of the cluster anymore!");
+    } else {
+      AttributeSet attributeSet = repository.attributeSetOf(request.attributeSetName());
+      if (attributeSet.isDefined()) {
+        attributeSet = AttributeSet.named(request.attributeSetName());
+        repository.remove(request.attributeSetName());
+      }
+      confirmingDistributor.confirmRemove(request.trackingId(), attributeSet, sourceNode);
     }
-    confirmingDistributor.confirmRemove(request.trackingId(), attributeSet, configuration.nodeMatching(request.sourceNodeId()));
   }
 
   void replaceAttribute(final ReceivedAttributeMessage request) {
     final AttributeSet attributeSet = repository.attributeSetOf(request.attributeSetName());
     if (attributeSet.isDefined()) {
-      final TrackedAttribute tracked = attributeSet.replace(request.attribute());
-      if (tracked.isPresent()) { // was both present and replaced
-        confirmingDistributor.confirm(request.trackingId(), attributeSet, tracked, request.type(), configuration.nodeMatching(request.sourceNodeId()));
+      final Node sourceNode = nodeLookup.apply(request.sourceNodeId());
+      if (sourceNode == null) {
+        logger.warn("Failed to replaceAttribute because source node " + request.sourceNodeId() + " is not part of the cluster anymore!");
+      } else {
+        final TrackedAttribute tracked = attributeSet.replace(request.attribute());
+        if (tracked.isPresent()) { // was both present and replaced
+          confirmingDistributor.confirm(request.trackingId(), attributeSet, tracked, request.type(), sourceNode);
+        }
       }
     }
   }
@@ -62,9 +89,14 @@ final class RemoteAttributeRequestHandler {
   void removeAttribute(final ReceivedAttributeMessage request) {
     final AttributeSet attributeSet = repository.attributeSetOf(request.attributeSetName());
     if (attributeSet.isDefined()) {
-      final TrackedAttribute tracked = attributeSet.remove(request.attribute());
-      if (tracked.isPresent()) { // actually was present, now removed
-        confirmingDistributor.confirm(request.trackingId(), attributeSet, tracked, request.type(), configuration.nodeMatching(request.sourceNodeId()));
+      final Node sourceNode = nodeLookup.apply(request.sourceNodeId());
+      if (sourceNode == null) {
+        logger.warn("Failed to removeAttribute because source node " + request.sourceNodeId() + " is not part of the cluster anymore!");
+      } else {
+        final TrackedAttribute tracked = attributeSet.remove(request.attribute());
+        if (tracked.isPresent()) { // actually was present, now removed
+          confirmingDistributor.confirm(request.trackingId(), attributeSet, tracked, request.type(), sourceNode);
+        }
       }
     }
   }
